==> analysis/0_anonymize.Rmd <==
---
title: "anonymize"
output: html_document
---


```{r}
library(dplyr)
library(ids) 
```

```{r}
adults_raw <- read.csv("/Users/karlaperez/Documents/projects/self-disclosure/data/DO_NOT_SHARE/MASTER_DATA_LONG.csv", 
                        stringsAsFactors = FALSE)

children_raw <- read.csv("~/Documents/projects/self-disclosure/data/DO_NOT_SHARE/data_children.csv", 
                       stringsAsFactors = FALSE)
```



```{r adults}

set.seed(12626)

## identify unique IDs
unique_ids_adults <- unique(adults_raw$prolific_id)

# create a mapping table
id_lookup_adults <- data.frame(
  prolific_id = unique_ids_adults,
  # generate random hex strings e.g., "a1b2c3d4"
  anon_id = random_id(length(unique_ids_adults), bytes = 4) 
)

# join back to the original data
anonymized_adults <- adults_raw %>%
  left_join(id_lookup_adults, by = "prolific_id") %>%
  select(-prolific_id, -source_file) %>%
  relocate(anon_id, .before = trial_index)

# write.csv(x = anonymized_adults, file = "anonymized_adults.csv", row.names = FALSE)

```


```{r children}

set.seed(12626)

## identify unique IDs
unique_ids_children <- unique(children_raw$subject_id)

# create a mapping table
id_lookup_children <- data.frame(
  subject_id = unique_ids_children,
  # generate random hex strings e.g., "a1b2c3d4"
  anon_id = random_id(length(unique_ids_children), bytes = 4) 
)

# join back to the original data
anonymized_df_children <- children_raw %>%
  left_join(id_lookup, by = "subject_id") %>%
  select(-subject_id) %>%
  relocate(anon_id, .before = age_at_test)

# write.csv(x = anonymized_df_children, file = "anonymized_children.csv", row.names = FALSE)

```
