==> analysis/0_demographics.Rmd <==
---
title: "demographics"
output: html_document
---

```{r}
library(dplyr)
```

```{r}
adult_demographics <- read.csv("~/Documents/projects/self-disclosure/data/DO_NOT_SHARE/prolific_demographics.csv")

MASTER_DATA_LONG <- read_csv("~/Documents/projects/self-disclosure/data/DO_NOT_SHARE/MASTER_DATA_LONG.csv")

anonymized_children <- read.csv("~/Documents/projects/self-disclosure/data/anonymized_children.csv", 
                           stringsAsFactors = FALSE)

children_ages <- anonymized_children %>%
  select(anon_id, age_at_test) %>% 
  rename(subject_id = anon_id)

```

```{r who is missing}

# all adults who completed the study (n = 252)
all_adults_demo <- adult_demographics %>% 
  filter(Status == "APPROVED") %>%
  pull(Participant.id)
  
# adults who actually made it to the final dataset (n = 250)
participant_adults_ids <- MASTER_DATA_LONG %>%
  distinct(prolific_id) %>% 
  pull(prolific_id)
  

setdiff(all_adults_demo, participant_adults_ids)

# [1] "5d22526f851927001e3a515a" "558823c2fdf99b318cb4224b"
# these two subjects completed the study but are missing from the final data-set

"5d22526f851927001e3a515a" %in% all_adults_demo # TRUE
"5d22526f851927001e3a515a" %in% participant_adults_ids # FALSE

"558823c2fdf99b318cb4224b" %in% all_adults_demo # TRUE
"558823c2fdf99b318cb4224b" %in% participant_adults_ids # FALSE


```


```{r final demographic calculation}

age_sex_adults <- adult_demographics %>% 
  filter(Participant.id %in% participant_adults_ids) %>% 
  select(Age, Sex) %>% 
  mutate(Age = as.numeric(Age), Sex = as.factor(Sex))

age_sex_adults %>% 
  group_by(Sex) %>% 
  count()

# 0.544% Female
# 0.452% Male
# 0.452 Prefer not to say
```

```{r}
age_sex_adults %>% 
  summarise(mean_age = mean(Age), sd_age = sd(Age), range(Age)) 
```
```{r}

# this child did not provide any responses
REMOVE_CHILDREN <- "aa35da33"

children_ages %>% 
  filter(!subject_id %in% REMOVE_CHILDREN) %>% 
  summarise(mean_age = mean(age_at_test), sd(age_at_test), range(age_at_test))

```
