==> analysis/1_preprocessing.Rmd <==
---
title: "preprocessing"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(stringr)
# library(text)
# library(here)
# library(reticulate)
# library(umap)
# library(viridis)
# library(plotly)
# library(purrr)
# library(lsr)
# 
# # # install python environment 
# # textrpp_install()
# 
# # initialize environment 
# textrpp_initialize(save_profile = TRUE)

```

```{r import data}

adults_raw <- read.csv("~/Documents/projects/self-disclosure/data/anonymized_adults.csv",
                       stringsAsFactors = FALSE)

children_raw <- read.csv("~/Documents/projects/self-disclosure/data/anonymized_children.csv", 
                       stringsAsFactors = FALSE)

any(unique(children_raw$anon_id) %in% unique(adults_raw$anon_id))

```

```{r clean data}

# adults' self disclosure trials -----------------------------------------------
adults_self <- adults_raw %>% 
  filter(trial_type == "survey-text") %>% 
  select(anon_id, trial_index, response, inter_item_latency_ms)

message(paste("total adult_self rows to embed:", nrow(adults_self))) # 3000


# adults' animal fluency trials  -----------------------------------------------
adults_animals <- adults_raw %>% 
  filter(trial_type == "animal_entry")

# children's self disclosure trials --------------------------------------------
children_self <- children_raw %>% 
  pivot_longer(cols = starts_with("utterance_"), 
               names_to = "utterance",
               names_prefix = "utterance_",
               values_to = "response")

message(paste("total child rows to embed:", nrow(children_self))) # 228

```


```{r standardize responses}

# all-distilroberta-v1 is case-sensitive, so we standardized all 3282 responses

preprocess_text <- function(text_column) {
  
  # replace NAs with "none"
  text_column <- ifelse(is.na(text_column), "none", text_column)
  
  text_column %>%
    tolower() %>%
    str_replace_all("[\r\n]", " ") %>%                 # remove newlines
    str_replace_all("[^[:alnum:][:space:]'.]", "") %>% # remove all punctuation EXCEPT apostrophes and periods
    str_squish() %>%                                   # remove whitespace
    {ifelse(. == "" | is.na(.), "none", .)}            # if response is "" or just spaces, make it "none"
}

adults_self$response_clean <- preprocess_text(adults_self$response)
children_self$response_clean <- preprocess_text(children_self$response)


# combine adult and child data for embeddings --------------------------------

## includes latency between responses; not used in this analysis 
adults_self <- adults_self %>%
  group_by(anon_id) %>%
  mutate(trial_index = row_number()) %>%
  ungroup()
  
adults_toCombine <- adults_self %>% 
  select(anon_id, trial_index, response_clean) %>% 
  mutate(age_group = "adult") %>% 
  rename(utterance = trial_index)
  
children_toCombine <- children_self %>% 
  select(anon_id, utterance, response_clean) %>%
  mutate(age_group = "child") 

all_responses_clean <- rbind(adults_toCombine, children_toCombine)
# write.csv(x = all_responses_clean, file = "all_responses_clean.csv", row.names = FALSE)

```


```{r}

# --- STEP 2: EMBEDDING (Optional: only if re-running) ---
# If you re-run, use clean_text. If not, use your existing child_vectors.
# child_vectors <- textEmbed(child_self$clean_text, model = "all-distilroberta-v1")

# --- STEP 3: THE DATA BRIDGE ---
child_data_long <- cbind(
  child_self[, c("subject_id", "age_at_test", "clean_text")], 
  as.data.frame(child_vectors)
) %>%
  mutate(
    is_skip = clean_text == "" | clean_text == "none",
    word_count = str_count(clean_text, "\\w+")
  )

# --- STEP 4: MATURITY (UTTERANCE LEVEL) ---
child_mat <- as.matrix(child_data_long[, 4:771])
child_norm <- child_mat / sqrt(rowSums(child_mat^2))
adult_centroid_norm <- adult_centroid / sqrt(sum(adult_centroid^2))

child_data_long$adult_typicality <- as.vector(child_norm %*% t(adult_centroid_norm))

# --- STEP 5: DIVERSITY (SUBJECT LEVEL) ---
child_metrics <- child_data_long %>%
  group_by(subject_id) %>%
  group_split() %>%
  map_dfr(function(df_sub) {
    real_data <- df_sub %>% filter(!is_skip)
    div_score <- NA
    if(nrow(real_data) >= 2) {
      v <- as.matrix(real_data[, 4:771])
      n <- v / sqrt(rowSums(v^2))
      sim_matrix <- n %*% t(n)
      diag(sim_matrix) <- NA
      div_score <- 1 - mean(sim_matrix, na.rm = TRUE)
    }
    data.frame(
      subject_id = df_sub$subject_id[1],
      age = first(df_sub$age_at_test),
      semantic_diversity = div_score,
      mean_typicality = mean(real_data$adult_typicality, na.rm = TRUE),
      mean_word_count = mean(real_data$word_count, na.rm = TRUE)
    )
  })

# --- STEP 6: THE FINAL MODELS ---
# These are the numbers for your CogSci table
summary(lm(semantic_diversity ~ mean_word_count + age, data = child_metrics))
summary(lm(mean_typicality ~ mean_word_count + age, data = child_metrics))
```
